# Reduce allocation in discovery by ~30%